Se agrgan los cors para la nueva pagina

+++ src/main/java/com/Store/Store/domain/dto/Product.java
@@ -22,7 +22,7 @@ public class Product {
     private String description;
 
     @Column(name = "status")
-    private byte status;
+    private int status;
 
     @Column(name = "stock")
     private long stock;
